requirements.txt
# ==========================================
# LangChain AI 助手 - Python 依赖包
# 适用于 Linux / macOS / Windows
# ==========================================

# Web 框架
streamlit

# LangChain 核心 (会自动安装 langchain-core)
langgraph
langchain-community

# Persistent Checkpoint (PostgreSQL)
langgraph-checkpoint-postgres
psycopg-pool
psycopg[binary]

# Google Gemini AI
langchain-google-genai
google-genai

# 向量数据库 (Qdrant)
langchain-qdrant

# 搜索工具
duckduckgo_search
ddgs

# Gmail 工具
langchain-google-community
google-auth-oauthlib

# 环境变量管理
python-dotenv
